Issue #8: Change language default value to None
This changes the default value of the language parameter in analyse_text(..) and analyse_html(..) to None instead of a fixed language.
This solves the issue that we force a default language from client side with a default language.
Further, we get the intended behavior that this results in the language being omitted as a API request parameter,
which enables the automatic language detection, if available in the pipeline.